feat: コンポーネントの設計変更
